completed RC driver with modifications

+++ stm32l552xx/boardfiles/drivers/rc/rc_defines.hpp
@@ -2,7 +2,7 @@
 #include <cstdint>
 
 /* define range mapping here, -+100% -> 1000..2000 */
-static constexpr uint8_t SBUS_RANGE_MIN = 192;
+static constexpr uint16_t SBUS_RANGE_MIN = 192;
 static constexpr uint16_t SBUS_RANGE_MAX = 1792;
 static constexpr uint16_t SBUS_RANGE_RANGE = (SBUS_RANGE_MAX - SBUS_RANGE_MIN);
